+++ test/models/folio/media_source_test.rb
@@ -75,6 +75,45 @@ class Folio::MediaSourceTest < ActiveSupport::TestCase
     assert_equal [oldest_link, newest_link], media_source.reload.media_source_site_links.to_a
   end
 
+  test "nested site rules can destroy and recreate the same site rule" do
+    media_source = create(:folio_media_source)
+    site = create(:dummy_site)
+    existing_link = media_source.media_source_site_links.create!(site:, max_usage_count: 20)
+
+    media_source.assign_attributes(
+      media_source_site_links_attributes: {
+        "0" => {
+          "id" => existing_link.id.to_s,
+          "site_id" => site.id.to_s,
+          "max_usage_count" => existing_link.max_usage_count.to_s,
+          "_destroy" => "1"
+        },
+        "1" => {
+          "id" => "",
+          "site_id" => site.id.to_s,
+          "max_usage_count" => "7",
+          "_destroy" => ""
+        }
+      }
+    )
+
+    assert media_source.save
+    assert_not Folio::MediaSourceSiteLink.exists?(existing_link.id)
+    assert_equal [site.id], media_source.reload.media_source_site_links.pluck(:site_id)
+    assert_equal 7, media_source.media_source_site_links.first.max_usage_count
+  end
+
+  test "site rules reject active duplicates" do
+    media_source = create(:folio_media_source)
+    site = create(:dummy_site)
+    media_source.media_source_site_links.create!(site:)
+
+    duplicate = media_source.media_source_site_links.build(site:)
+
+    assert_not duplicate.valid?
+    assert_includes duplicate.errors.details[:media_source_id].map { |e| e[:error] }, :taken
+  end
+
   test "destroy nullifies attached files attributes and removes only media-source sites" do
     media_source = create(:folio_media_source)
     site1 = create(:folio_site, domain: "site1.localhost", type: "Folio::Site")